Conflicts:
	cmd/mount_zfs/mount_zfs.c
	cmd/zfs/zfs_main.c
	cmd/zfs_util/zfs_util.c
	include/sys/zvolIO.h

+++ module/zfs/zio_inject.c
@@ -348,7 +348,7 @@ spa_handle_ignored_writes(spa_t *spa)
 			    ddi_time_after64(
 			    (int64_t)handler->zi_record.zi_timer +
 			    handler->zi_record.zi_duration * hz,
-			    ddi_get_lbolt64()));
+			    (int64_t)ddi_get_lbolt64()));
 		} else {
 			/* duration is negative so the subtraction here adds */
 			VERIFY(handler->zi_record.zi_timer == 0 ||
